update en aanpassing fields lijsten
- update op evt. wijzigingen aan specificaties sinds vorige keer van genereren
- doorvoeren wijzigingen in fields logica:
    - automatisch leveren rni en verificatie
    - altijd leveren hele waardetabel
    - verkort pad is niet meer mogelijk

gevolgen:
- geen mapping van verkort pad naar volledig pad relevant
- geen lijst met kortste pad meer relevant
- wel lijst met alle mogelijke/toegestane paden plus lijst met alleen relevante paden (excl. automatisch geleverde velden)
