+++ features/fields-PersoonBeperkt.csv
@@ -1,30 +1,96 @@
-fields pad,volledig pad
-adressering,adressering
-adresregel1,adressering.adresregel1
-adresregel2,adressering.adresregel2
-adresregel3,adressering.adresregel3
-adressering.land,adressering.land
-burgerservicenummer,burgerservicenummer
-geboorte,geboorte
-geboorte.datum,geboorte.datum
-geslacht,geslacht
-leeftijd,leeftijd
-naam,naam
-naam.adellijkeTitelPredicaat,naam.adellijkeTitelPredicaat
-naam.geslachtsnaam,naam.geslachtsnaam
-volledigeNaam,naam.volledigeNaam
-naam.voorletters,naam.voorletters
-naam.voornamen,naam.voornamen
-naam.voorvoegsel,naam.voorvoegsel
-opschortingBijhouding,opschortingBijhouding
-overlijden,overlijden
-indicatieOverleden,overlijden.indicatieOverleden
-rni,rni
-categorie,rni.categorie
-deelnemer,rni.deelnemer
-omschrijvingVerdrag,rni.omschrijvingVerdrag
-verblijfplaats,verblijfplaats
-functieAdres,verblijfplaats.functieAdres
-verificatie,verificatie
-verificatie.datum,verificatie.datum
-verificatie.omschrijving,verificatie.omschrijving
+adressering
+adressering.adresregel1
+adressering.adresregel2
+adressering.adresregel3
+adressering.land
+adressering.land.code
+adressering.land.omschrijving
+adressering.inOnderzoek
+adressering.inOnderzoek.adresregel1
+adressering.inOnderzoek.adresregel2
+adressering.inOnderzoek.adresregel3
+adressering.inOnderzoek.datumIngangOnderzoekVerblijfplaats
+adressering.inOnderzoek.datumIngangOnderzoekVerblijfplaats.langFormaat
+adressering.inOnderzoek.datumIngangOnderzoekVerblijfplaats.type
+adressering.inOnderzoek.datumIngangOnderzoekVerblijfplaats.datum
+adressering.inOnderzoek.datumIngangOnderzoekVerblijfplaats.onbekend
+adressering.inOnderzoek.datumIngangOnderzoekVerblijfplaats.jaar
+adressering.inOnderzoek.datumIngangOnderzoekVerblijfplaats.maand
+adressering.inOnderzoek.land
+burgerservicenummer
+geboorte
+geboorte.datum
+geboorte.datum.langFormaat
+geboorte.datum.type
+geboorte.datum.datum
+geboorte.datum.onbekend
+geboorte.datum.jaar
+geboorte.datum.maand
+geboorte.inOnderzoek
+geboorte.inOnderzoek.datumIngangOnderzoek
+geboorte.inOnderzoek.datumIngangOnderzoek.langFormaat
+geboorte.inOnderzoek.datumIngangOnderzoek.type
+geboorte.inOnderzoek.datumIngangOnderzoek.datum
+geboorte.inOnderzoek.datumIngangOnderzoek.onbekend
+geboorte.inOnderzoek.datumIngangOnderzoek.jaar
+geboorte.inOnderzoek.datumIngangOnderzoek.maand
+geboorte.inOnderzoek.datum
+geheimhoudingPersoonsgegevens
+geslacht
+geslacht.code
+geslacht.omschrijving
+inOnderzoek
+inOnderzoek.burgerservicenummer
+inOnderzoek.datumIngangOnderzoekPersoon
+inOnderzoek.datumIngangOnderzoekPersoon.langFormaat
+inOnderzoek.datumIngangOnderzoekPersoon.type
+inOnderzoek.datumIngangOnderzoekPersoon.datum
+inOnderzoek.datumIngangOnderzoekPersoon.onbekend
+inOnderzoek.datumIngangOnderzoekPersoon.jaar
+inOnderzoek.datumIngangOnderzoekPersoon.maand
+inOnderzoek.geslacht
+inOnderzoek.leeftijd
+leeftijd
+naam
+naam.adellijkeTitelPredicaat
+naam.adellijkeTitelPredicaat.code
+naam.adellijkeTitelPredicaat.omschrijving
+naam.adellijkeTitelPredicaat.soort
+naam.geslachtsnaam
+naam.volledigeNaam
+naam.voorletters
+naam.voornamen
+naam.voorvoegsel
+naam.inOnderzoek
+naam.inOnderzoek.datumIngangOnderzoek
+naam.inOnderzoek.datumIngangOnderzoek.langFormaat
+naam.inOnderzoek.datumIngangOnderzoek.type
+naam.inOnderzoek.datumIngangOnderzoek.datum
+naam.inOnderzoek.datumIngangOnderzoek.onbekend
+naam.inOnderzoek.datumIngangOnderzoek.jaar
+naam.inOnderzoek.datumIngangOnderzoek.maand
+naam.inOnderzoek.adellijkeTitelPredicaat
+naam.inOnderzoek.geslachtsnaam
+naam.inOnderzoek.voorletters
+naam.inOnderzoek.voornamen
+naam.inOnderzoek.voorvoegsel
+naam.inOnderzoek.volledigeNaam
+opschortingBijhouding
+opschortingBijhouding.reden
+opschortingBijhouding.reden.code
+opschortingBijhouding.reden.omschrijving
+rni
+rni.categorie
+rni.deelnemer
+rni.deelnemer.code
+rni.deelnemer.omschrijving
+rni.omschrijvingVerdrag
+verificatie
+verificatie.datum
+verificatie.datum.langFormaat
+verificatie.datum.type
+verificatie.datum.datum
+verificatie.datum.onbekend
+verificatie.datum.jaar
+verificatie.datum.maand
+verificatie.omschrijving
