+++ features/fields-filtered-PersoonBeperkt.csv
@@ -1,0 +1,19 @@
+adressering
+adressering.adresregel1
+adressering.adresregel2
+adressering.adresregel3
+adressering.land
+burgerservicenummer
+geboorte
+geboorte.datum
+geslacht
+leeftijd
+naam
+naam.adellijkeTitelPredicaat
+naam.adellijkeTitelPredicaat.soort
+naam.geslachtsnaam
+naam.volledigeNaam
+naam.voorletters
+naam.voornamen
+naam.voorvoegsel
+opschortingBijhouding
