+++ features/fields-generator/brp_persoon.json
@@ -16,18 +16,17 @@
         "type",
         "inOnderzoek",
         "geheimhoudingPersoonsgegevens",
-        "opschortingBijhouding.reden",
-        "jaar",
-        "maand",
-        "onbekend",
-        "langFormaat"
+        "reden",
+        "rni",
+        "verificatie"
     ],
-    "dateComponents": [
+    "autoComponents": [
         "AbstractDatum",
         "VolledigeDatum",
         "DatumOnbekend",
         "JaarDatum",
-        "JaarMaandDatum"
+        "JaarMaandDatum",
+        "Waardetabel"
     ],
     "sourceYaml": "/specificatie/genereervariant/openapi.yaml",
     "fieldslistFolder": "/features/"
