+++ features/fields-generator/readme.md
@@ -1,35 +1,44 @@
 # Fields generator
 
-Deze map bevat Python 3 scripts waarmee vanuit de open API specificaties de volgende bestanden worden gegenereerd:
-- fields-mapping-Persoon.csv
-- fields-mapping-PersoonBeperkt.csv
+Deze map bevat Python een script waarmee vanuit de open API specificaties de volgende bestanden worden gegenereerd:
 - fields-persoon.csv
 - fields-persoonBeperkt.csv
+- fields-filtered-persoon.csv
+- fields-filtered-persoonBeperkt.csv
+
+## fields-Persoon en fields-PersoonBeperkt
+De bestanden fields-Persoon.csv en fields-PersoonBeperkt.csv bevatten alle mogelijke (toegestane) paden die in de "fields" parameter ingevuld mogen worden.
+
+Deze lijst bevat ook velden die niet expliciet met fields gevraagd hoeven te worden, omdat die al automatisch geleverd zullen worden (bijv. geheimhouding, inOnderzoek en type). Deze mag je in fields invullen, maar worden genegeerd.
+
+Deze lijst bevat ook velden van een datum en waardetabel. Die mag je invullen, maar in dat geval worden alle velden van dat datumveld dan wel die waardetabel (code en omschrijving) geleverd.
+
+Er zijn rijen opgenomen voor losse velden (simple type velden zoals string, integer, boolean), en ook voor het vragen van hele groepen (objecten).
+
+## fields-filtered-Persoon en fields-filtered-PersoonBeperkt
+De bestanden fields-filtered-Persoon.csv en fields-filtered-PersoonBeperkt.csv bevatten alle paden die in de "fields" parameter ingevuld kunnen worden.
+
+Deze lijst bevat geen velden die niet expliciet met fields gevraagd hoeven te worden, omdat die al automatisch geleverd zullen worden (bijv. geheimhouding, inOnderzoek en type). 
+Deze lijst bevat ook geen sub-velden van een datum of waardetabel, omdat datum- of waardetabelobjecten altijd in zijn geheel geleverd worden.
+
+Er zijn rijen opgenomen voor losse velden (simple type velden zoals string, integer, boolean), en ook voor het vragen van hele groepen (objecten).
 
 ## Voor je begint met genereren
 - Voor Persoon en PersoonBeperkt gebruiken we brp_persoon.json. In de configuratie moet je in ieder geval "projectFolder" aanpassen, zodat dit correct verwijst naar waar de GitHub repository van BRP Personen bevragen staat, met daarin de specificaties (/specificatie/genereervariant/openapi.yaml) en de locatie waar de gegenereerde bestanden geplaatst worden (/features).
 - Installeer -als je dat nog niet hebt- Python 3. De genereerscripts werken mogelijk niet correct met python 2.
 - Installeer - als je dat nog niet hebt- het pyyaml package.
 
-## fields-mapping-Persoon en fields-mapping-PersoonBeperkt
-De bestanden fields-mapping-Persoon.csv en fields-mapping-PersoonBeperkt.csv bevatten alle mogelijke (toegestane) paden die verwijzen naar een veld in de response. Naar één veld kan op verschillende manieren worden verwezen, door een verkort pad te gebruiken. Er zijn rijen per veld, maar ook rijen voor het opvragen van een hele groep. 
-Bij elk pad staat het bijbehorende volledige pad dat naar het veld verwijst. Bijvoorbeeld aanschrijfwijze,adressering.aanschrijfwijze. Dit betekent dat fields=["aanschrijfwijze"] het veld adressering.aanschrijfwijze moet leveren.
-
-De bestanden fields-mapping-Persoon.csv en fields-mapping-PersoonBeperkt.csv worden gegenereerd door het script fields_mapping_generator.py. 
+## Genereren van de fields lijsten
+De bestanden worden gegenereerd door het script fields_generator.py. 
 
 Dit script vraagt minimaal command line argument --config met daarin een verwijzing naar het configuratiebestand. 
 
 Voer in de command line uit met bijvoorbeeld:
-`python3 fields_mapping_generator.py --config"brp_persoon.json" --info`
+`python3 fields_generator.py --config"brp_persoon.json"`
 
-## fields-persoon en fields-persoonBeperkt
-De bestanden fields-persoon.csv en fields-persoonBeperkt.csv bevatten per veld in de response één rij met het kortst mogelijke pad dat uniek verwijst naar het veld.
+Om fields-filtered-Persoon en fields-filtered-PersoonBeperkt te genereren voeg je command line argument --filter toe.
+Bijvoorbeeld:
+`python3 fields_generator.py --config"brp_persoon.json" --filter`
 
-De bestanden fields-persoon en fields-persoonBeperkt worden gegenereerd door het script fields_generator.py. 
 
-Dit script gebruikt fields-mapping-Persoon.csv en fields-mapping-PersoonBeperkt.csv als input, dus fields mapping moet -na bijvoorbeeld specificatiewijzigingen- eerst gedraaid worden voor fields_generator.py kan worden gedraaid. De scripts in omgekeerde volgorde uitvoeren kan leiden tot inconsistenties tussen de verschillende bestanden.
 
-Dit script vraagt minimaal command line argument --config met daarin een verwijzing naar het configuratiebestand. 
-
-Voer in de command line uit met bijvoorbeeld:
-`python3 fields_generator.py --config"brp_persoon.json" --info`
