--- features/fields-mapping-PersoonBeperkt.csv
@@ -1,86 +1,0 @@
-fields pad,volledig pad
-adressering,adressering
-adressering.adresregel1,adressering.adresregel1
-adresregel1,adressering.adresregel1
-adressering.adresregel2,adressering.adresregel2
-adresregel2,adressering.adresregel2
-adressering.adresregel3,adressering.adresregel3
-adresregel3,adressering.adresregel3
-adressering.land,adressering.land
-land,adressering.land
-adressering.land.code,adressering.land
-land.code,adressering.land
-adressering.land.omschrijving,adressering.land
-land.omschrijving,adressering.land
-burgerservicenummer,burgerservicenummer
-geboorte,geboorte
-geboorte.datum,geboorte.datum
-geboorte.datum.langFormaat,geboorte.datum
-geboorte.datum.type,geboorte.datum
-geboorte.datum.datum,geboorte.datum
-geboorte.datum.onbekend,geboorte.datum
-geboorte.datum.jaar,geboorte.datum
-geboorte.datum.maand,geboorte.datum
-geheimhoudingPersoonsgegevens,geheimhoudingPersoonsgegevens
-geslacht,geslacht
-geslacht.code,geslacht
-geslacht.omschrijving,geslacht
-leeftijd,leeftijd
-naam,naam
-naam.adellijkeTitelPredicaat,naam.adellijkeTitelPredicaat
-adellijkeTitelPredicaat,naam.adellijkeTitelPredicaat
-naam.adellijkeTitelPredicaat.code,naam.adellijkeTitelPredicaat
-adellijkeTitelPredicaat.code,naam.adellijkeTitelPredicaat
-naam.adellijkeTitelPredicaat.omschrijving,naam.adellijkeTitelPredicaat
-adellijkeTitelPredicaat.omschrijving,naam.adellijkeTitelPredicaat
-naam.adellijkeTitelPredicaat.soort,naam.adellijkeTitelPredicaat
-adellijkeTitelPredicaat.soort,naam.adellijkeTitelPredicaat
-soort,naam.adellijkeTitelPredicaat
-naam.geslachtsnaam,naam.geslachtsnaam
-geslachtsnaam,naam.geslachtsnaam
-naam.volledigeNaam,naam.volledigeNaam
-volledigeNaam,naam.volledigeNaam
-naam.voorletters,naam.voorletters
-voorletters,naam.voorletters
-naam.voornamen,naam.voornamen
-voornamen,naam.voornamen
-naam.voorvoegsel,naam.voorvoegsel
-voorvoegsel,naam.voorvoegsel
-opschortingBijhouding,opschortingBijhouding
-opschortingBijhouding.reden,opschortingBijhouding.reden
-reden,opschortingBijhouding.reden
-opschortingBijhouding.reden.code,opschortingBijhouding.reden
-reden.code,opschortingBijhouding.reden
-opschortingBijhouding.reden.omschrijving,opschortingBijhouding.reden
-reden.omschrijving,opschortingBijhouding.reden
-overlijden,overlijden
-overlijden.indicatieOverleden,overlijden.indicatieOverleden
-indicatieOverleden,overlijden.indicatieOverleden
-rni,rni
-rni.categorie,rni.categorie
-categorie,rni.categorie
-rni.deelnemer,rni.deelnemer
-deelnemer,rni.deelnemer
-rni.deelnemer.code,rni.deelnemer
-deelnemer.code,rni.deelnemer
-rni.deelnemer.omschrijving,rni.deelnemer
-deelnemer.omschrijving,rni.deelnemer
-rni.omschrijvingVerdrag,rni.omschrijvingVerdrag
-omschrijvingVerdrag,rni.omschrijvingVerdrag
-verblijfplaats,verblijfplaats
-verblijfplaats.type,verblijfplaats.type
-verblijfplaats.functieAdres,verblijfplaats.functieAdres
-functieAdres,verblijfplaats.functieAdres
-verblijfplaats.functieAdres.code,verblijfplaats.functieAdres
-functieAdres.code,verblijfplaats.functieAdres
-verblijfplaats.functieAdres.omschrijving,verblijfplaats.functieAdres
-functieAdres.omschrijving,verblijfplaats.functieAdres
-verificatie,verificatie
-verificatie.datum,verificatie.datum
-verificatie.datum.langFormaat,verificatie.datum
-verificatie.datum.type,verificatie.datum
-verificatie.datum.datum,verificatie.datum
-verificatie.datum.onbekend,verificatie.datum
-verificatie.datum.jaar,verificatie.datum
-verificatie.datum.maand,verificatie.datum
-verificatie.omschrijving,verificatie.omschrijving
